sharing: Set appropriate filename of shared content
previously filename was the last element of sharedDataUrl.split('/'),
this did not always correspond to the actual filename, sometimes
this element did not contain a file extension causing the uploaded
data to be stored as a binary file. this made shared images fail to
render and on downloading the shared content it was downloaded as
a binary file, this commit gives the files generic names with
appropriate extensions.